Wow, awesome new database interface. Not compile errors, but I haven't run tests. Committing w.i.p.

+++ server/src/main/java/nz/ac/vuw/ecs/rprofs/server/domain/id/DataSetId.java
@@ -8,7 +8,7 @@ import nz.ac.vuw.ecs.rprofs.server.domain.Dataset;
  */
 public class DatasetId extends Id<DatasetId, Dataset> {
 
-	public DatasetId(short id) {
+	public DatasetId(long id) {
 		super(id);
 	}
